% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureSummary.R
\docType{methods}
\name{addMarginalBatch}
\alias{addMarginalBatch}
\alias{addMarginalBatch,MixtureSummary-method}
\title{TODO add title}
\usage{
addMarginalBatch(summ)

\S4method{addMarginalBatch}{MixtureSummary}(summ)
}
\arguments{
\item{summ}{see \code{showMethods(addMarginalBatch)}}
}
\value{
A `MixtureSummary` object
}
\description{
TODO document
}
