% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureSummary.R,
%   R/methods-CopyNumberMixtureSummary.R
\docType{methods}
\name{addMarginalModel}
\alias{addMarginalModel}
\alias{addMarginalModel,MixtureSummary-method}
\alias{addMarginalModel,CopyNumberMixtureSummary-method}
\title{TODO add title}
\usage{
addMarginalModel(summ)

\S4method{addMarginalModel}{MixtureSummary}(summ)

\S4method{addMarginalModel}{CopyNumberMixtureSummary}(summ)
}
\arguments{
\item{summ}{see \code{showMethods(addMarginalModel)}}
}
\value{
A `MixtureSummary` object
}
\description{
TODO document
}
