% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CNPBayes_MBP}
\alias{CNPBayes_MBP}
\title{This data is an instance of MultiBatchPooled}
\format{An object of class \code{MultiBatchPooled} of length 1.}
\usage{
CNPBayes_MBP
}
\value{
An example of a `MultiBatchPooled`
CNPBayes_MBP
}
\description{
This data is an instance of MultiBatchPooled
}
\author{
David McKean
}
\keyword{datasets}
