% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CNPBayes_SBP}
\alias{CNPBayes_SBP}
\title{This data is an instance of SingleBatchPooled}
\format{An object of class \code{SingleBatchPooled} of length 1.}
\usage{
CNPBayes_SBP
}
\value{
An example of a `SingleBatchPooled`
CNPBayes_SBP
}
\description{
This data is an instance of SingleBatchPooled
}
\author{
David McKean
}
\keyword{datasets}
