% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{CopyNumberMixtureSummary-class}
\alias{CopyNumberMixtureSummary-class}
\title{A plottable summary of a mixture model and inferred copy number}
\description{
An object summarizing the observed mixture model and theoretical models as plottable data.frames.
}
\section{Slots}{

\describe{
\item{\code{observed}}{data.frame of observed values. Must have the following columns: "batch" (bactor), "component" (factor), "copynumber" (factor), "x.val" (numeric). Otherwise identical to `MixtureSummary` .}

\item{\code{predicted}}{data.frame of predicted densities. Must have the following columns: "batch" (factor), "component" (factor), "copynumber" (factor), "x" (numeric), "y" (numeric). Otherwise identical to `MixtureSummary` .}

\item{\code{nBins}}{integer value for reccomended number of bins when plotting over the exact range of `.Data@observed$x.val` . Identical to `MixtureSummary` .}
}}

