% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-GraphicalParameters.R
\docType{methods}
\name{getHistogramParams}
\alias{getHistogramParams}
\alias{getHistogramParams,GraphicalParameters-method}
\title{Get graphical parameters for plotting a histogram of observed values.}
\usage{
getHistogramParams(params)

\S4method{getHistogramParams}{GraphicalParameters}(params)
}
\arguments{
\item{params}{see \code{showMethods(k)}}
}
\value{
A named list of parameters that may be passed to a ggplot::geom_histogram function call.
}
\description{
This function retrieves the non-NULL values from histogram.* slots in a GraphicalParameters object.
}
