% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-GraphicalParameters.R
\docType{methods}
\name{getLineParams}
\alias{getLineParams}
\alias{getLineParams,GraphicalParameters-method}
\title{Get graphical parameters for plotting a density line of theoretical models.}
\usage{
getLineParams(params)

\S4method{getLineParams}{GraphicalParameters}(params)
}
\arguments{
\item{params}{see \code{showMethods(k)}}
}
\value{
A named list of parameters that may be passed to a ggplot::geom_line function call.
}
\description{
This function retrieves the non-NULL values from line.* slots in a GraphicalParameters object.
}
