% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureSummary.R
\docType{methods}
\name{getObserved}
\alias{getObserved}
\alias{getObserved,MixtureSummary-method}
\title{Observed model summary}
\usage{
getObserved(object)

\S4method{getObserved}{MixtureSummary}(object)
}
\arguments{
\item{object}{see \code{showMethods(getObserved)}}
}
\value{
A data.frame. Column names must include those returned by `ObservedNames(object)`.
}
\description{
This function retrieves a summary of the an observed model.
}
