% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{getSigmaMatrix}
\alias{getSigmaMatrix}
\alias{getSigmaMatrix,MixtureModel-method}
\alias{getSigmaMatrix,SingleBatchModel-method}
\alias{getSigmaMatrix,MultiBatchPooled-method}
\alias{getSigmaMatrix,MultiBatchCopyNumberPooled-method}
\title{Wrapper function for a `MixtureModel`'s sigma accessor.}
\usage{
getSigmaMatrix(model)

\S4method{getSigmaMatrix}{MixtureModel}(model)

\S4method{getSigmaMatrix}{SingleBatchModel}(model)

\S4method{getSigmaMatrix}{MultiBatchPooled}(model)

\S4method{getSigmaMatrix}{MultiBatchCopyNumberPooled}(model)
}
\arguments{
\item{object}{see \code{showMethods(k)}}
}
\value{
A "b X k" numeric matrix where b is the number batches and k is the number of predicted components.
}
\description{
This function retrieves the model's sigma values as a matrix.
}
