% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureSummary.R
\docType{methods}
\name{getTheoretical}
\alias{getTheoretical}
\alias{getTheoretical,MixtureSummary-method}
\title{Observed model summary}
\usage{
getTheoretical(object)

\S4method{getTheoretical}{MixtureSummary}(object)
}
\arguments{
\item{object}{see \code{showMethods(getTheoretical)}}
}
\value{
A data.frame. Column names must include those returned by `TheoreticalNames(object)`.
}
\description{
This function retrieves a summary of the a theoretical model.
}
