% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{getThetaMatrix}
\alias{getThetaMatrix}
\alias{getThetaMatrix,MixtureModel-method}
\alias{getThetaMatrix,SingleBatchModel-method}
\title{Wrapper function for a `MixtureModel`'s theta accessor.}
\usage{
getThetaMatrix(model)

\S4method{getThetaMatrix}{MixtureModel}(model)

\S4method{getThetaMatrix}{SingleBatchModel}(model)
}
\arguments{
\item{model}{see \code{showMethods(k)}}
}
\value{
A "b X k" numeric matrix where b is the number batches and k is the number of predicted components.
}
\description{
This function retrieves the model's theta values as a matrix.
}
