% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{MixModelViz}
\alias{MixModelViz}
\alias{MixModelViz-package}
\title{Plot mixture models with applications to copy number / batch effects}
\description{
Plot mixture models with applications to copy number / batch effects
}
