% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MixtureSummary-class}
\alias{MixtureSummary-class}
\title{A plottable summary of a mixture model}
\description{
An object summarizing the observed mixture model and theoretical models as plottable data.frames
}
\section{Slots}{

\describe{
\item{\code{observed}}{data.frame of observed values, with batch and component metadata. Suitable for histogram plotting. Must have the following columns: "batch" (bactor), "component" (factor), "x.val" (numeric).}

\item{\code{predicted}}{data.frame of predicted densities plotted over the range of observed values, with batch and component metadata. Suitable for line plotting. Must have the following columns: "batch" (factor), "component" (factor), "x" (numeric), "y" (numeric).}

\item{\code{nBins}}{integer value for reccomended number of bins when plotting over the exact range of `.Data@observed$x.val` . `.Data@predicted$y` has been scaled to bin width, e.g. `diff(range(.Data@observed$x.val))/nBins` .}
}}

