% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureSummary.R
\docType{methods}
\name{nBins}
\alias{nBins}
\alias{nBins,MixtureSummary-method}
\title{Number of bins}
\usage{
nBins(object)

\S4method{nBins}{MixtureSummary}(object)
}
\arguments{
\item{object}{see \code{showMethods(nBins)}}
}
\value{
integer
}
\description{
This function retrieves the number of bins suggested for plotting an observed model summary as a histogram over the range of it's x values.
}
