% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureSummary.R,
%   R/methods-CopyNumberMixtureSummary.R
\docType{methods}
\name{ObservedNames}
\alias{ObservedNames}
\alias{ObservedNames,MixtureSummary-method}
\alias{ObservedNames,CopyNumberMixtureSummary-method}
\title{Observed variables}
\usage{
ObservedNames(object)

\S4method{ObservedNames}{MixtureSummary}(object)

\S4method{ObservedNames}{CopyNumberMixtureSummary}(object)
}
\arguments{
\item{object}{see \code{showMethods(ObservedNames)}}
}
\value{
A character vector.
}
\description{
This function shows that column names that must be present in an observed model summary. These must be present in the character vector `names(getObserved(object))`. Checked by `validObject(object)`.
}
