% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{summarizeObserved}
\alias{summarizeObserved}
\alias{summarizeObserved,MixtureModel-method}
\alias{summarizeObserved,SingleBatchModel-method}
\alias{summarizeObserved,SingleBatchCopyNumber-method}
\alias{summarizeObserved,MultiBatchCopyNumber-method}
\alias{summarizeObserved,MultiBatchCopyNumberPooled-method}
\title{Create a plottable summary of the observed mixture model.}
\usage{
summarizeObserved(model)

\S4method{summarizeObserved}{MixtureModel}(model)

\S4method{summarizeObserved}{SingleBatchModel}(model)

\S4method{summarizeObserved}{SingleBatchCopyNumber}(model)

\S4method{summarizeObserved}{MultiBatchCopyNumber}(model)

\S4method{summarizeObserved}{MultiBatchCopyNumberPooled}(model)
}
\arguments{
\item{model}{A `MixtureModel` object}
}
\value{
A data.frame containing the varibles described by `ObservedNames(model)` for each observed object in the model.
}
\description{
This function retrieves the observed values, predicted component, and batch identifiers from a `MixtureModel` object. Suitable for plotting counts with a histogram.
}
