% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{summarizeTheoretical}
\alias{summarizeTheoretical}
\alias{summarizeTheoretical,MixtureModel,data.frame-method}
\alias{summarizeTheoretical,MixtureModel-method}
\alias{summarizeTheoretical,SingleBatchCopyNumber,data.frame-method}
\alias{summarizeTheoretical,SingleBatchCopyNumber-method}
\alias{summarizeTheoretical,MultiBatchCopyNumber,data.frame-method}
\alias{summarizeTheoretical,MultiBatchCopyNumber-method}
\alias{summarizeTheoretical,MultiBatchCopyNumberPooled,data.frame-method}
\alias{summarizeTheoretical,MultiBatchCopyNumberPooled-method}
\title{Create a plottable summary of the theoretical mixture model.}
\usage{
summarizeTheoretical(model, obs.df)

\S4method{summarizeTheoretical}{MixtureModel,data.frame}(model, obs.df)

\S4method{summarizeTheoretical}{SingleBatchCopyNumber,data.frame}(model, obs.df)

\S4method{summarizeTheoretical}{MultiBatchCopyNumber,data.frame}(model, obs.df)

\S4method{summarizeTheoretical}{MultiBatchCopyNumberPooled,data.frame}(model,
  obs.df)
}
\arguments{
\item{model}{A `MixtureModel` object}

\item{obs.df}{data.frame of observed values generated by `getObserved(model)`. Column names must include those returned by `ObservedNames(model)`.}
}
\value{
A data.frame containing the varibles described by `TheoreticalNames(model)` for each of the observed distributions.
}
\description{
This function retrieves the theoretical distribution from a `MixtureModel` object and generates density points over the range of x values given in the model's observed data. Suitable for plotting as a line. Y values have been scaled to fit the histogram over the range of the observed data when split using `nBins(MixtureSummary.obj)`.
}
