% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{plot_model}
\alias{plot_model}
\title{Plotting function for `MixtureSummary` objects}
\usage{
plot_model(summ, params)
}
\arguments{
\item{summ}{a `MixtureSummary` object.}

\item{params}{A GraphicalParamters object. If missing, a new object will be
instantiated.
GraphicalParameters includes a color palette for fill and color
aesthetics, which will be included via `scale_*_manual`. Setting these
palettes to NULL will cause the `scale_*_manual` function call to be
removed from the ggplot object.
GraphicalParameters also includes default alpha, linetype, and size
aesthetics for the geom_histogram and geom_line layers. These values
may be set by vectors of length 1 or of a length equal to number of
data points represented in these layers. Setting a value to NULL will
cause the aesthetic to be removed from the geom_* function call.}
}
\value{
An object of class `ggplot`
}
\description{
Plots normal densities from theoretical mixture models over a histogram of observed data.
}
\examples{
data(SingleBatchPooledExample, package="MixModelViz")
sbp.summ <- summarize(SingleBatchPooledExample)
plot_model(sbp.summ)

data(MultiBatchPooledExample, package="MixModelViz")
mbcnp.summ <- summarize(CopyNumberModel(MultiBatchPooledExample))
plot_model(mbcnp.summ, new("GraphicalParameters", line.size=2))

}
