% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Plotting function for `MixtureSummary` objects}
\usage{
plot_model(summ, fill_aes, palette = c(`1` = "#56B4E9", `2` = "#E69F00", `3` =
  "#009E73", `4` = "#F0E442", `5` = "#0072B2"), fixed_aes = list(hist =
  list(alpha = 0.5, linetype = 0, size = 0), line = list(alpha = 1, linetype =
  1, size = 1)))
}
\arguments{
\item{summ}{a `MixtureSummary` object}

\item{fill_aes}{should be "copynumber" or "component", but can be any value
that can be evaluated by `aes_(fill=as.name(fill_aes))` where `data=obs.df`}

\item{palette}{A vector of colors. To specify separate vectors for fill and}

\item{fixed_aes}{a list of alpha, linetype, and size values that will be
passed on to geom_hist and geom_line as fixed aesthetic parameters.
The named list should contain named vectors "hist" and "line".}
}
\value{
An object of class `ggplot`
}
\description{
Plots normal densities from theoretical mixture models over a histogram of observed data.
}
\examples{
data(SingleBatchPooledExample, package="MixModelViz")
sbp.summ <- summarize(SingleBatchPooledExample)
plot_model(sbp.summ)

data(MultiBatchPooledExample, package="MixModelViz")
mbcnp.summ <- summarize(CopyNumberModel(MultiBatchPooledExample))
plot_model(mbcnp.summ)
}
