% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureModel.R
\docType{methods}
\name{summarize}
\alias{summarize}
\alias{summarize,TempModel-method}
\alias{summarize,MixtureModel-method}
\title{Summarize a MixtureModel into a set of data.frames that can be plotted
 (observed and predicted)}
\usage{
summarize(model)

\S4method{summarize}{TempModel}(model)
}
\arguments{
\item{model}{a MixtureModel-derived object}
}
\value{
A MixtureSummary object of two data.frames (observed and predicted) that can be plotted
}
\description{
Summarize a MixtureModel into a set of data.frames that can be plotted
 (observed and predicted)
}
