% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MixtureSummary.R,
%   R/methods-CopyNumberMixtureSummary.R
\docType{methods}
\name{TheoreticalNames}
\alias{TheoreticalNames}
\alias{TheoreticalNames,MixtureSummary-method}
\alias{TheoreticalNames,CopyNumberMixtureSummary-method}
\title{Theoretical variables}
\usage{
TheoreticalNames(object)

\S4method{TheoreticalNames}{MixtureSummary}(object)

\S4method{TheoreticalNames}{CopyNumberMixtureSummary}(object)
}
\arguments{
\item{object}{see \code{showMethods(TheoreticalNames)}}
}
\value{
A character vector.
}
\description{
This function shows that column names that must be present in a theoretical model summary. These must be present in the character vector `names(getTheoretical(object))`. Checked by `validObject(object)`.
}
